Rename files to match module nesting
