+++ test/support/rails_application_fixture_helper.rb
@@ -1,7 +1,7 @@
 # typed: true
 # frozen_string_literal: true
 
-require "rails_test_helper"
+require "support/rails_test_helper"
 require "zeitwerk"
 
 module RailsApplicationFixtureHelper
