+++ test/unit/packwerk/constant_discovery_test.rb
