+++ test/unit/packwerk/formatters/offenses_formatter_test.rb
