+++ test/unit/packwerk/formatters/progress_formatter_test.rb
