+++ test/unit/packwerk/generators/configuration_file_test.rb
