+++ test/unit/packwerk/generators/root_package_test.rb
