+++ test/unit/packwerk/reference_extractor_test.rb
@@ -2,7 +2,7 @@
 # frozen_string_literal: true
 
 require "test_helper"
-require "parser_test_helper"
+require "support/packwerk/parser_test_helper"
 require "constant_resolver"
 
 module Packwerk
